+++ src/ff7/battle/animations.cpp
@@ -34,6 +34,7 @@
 #include "defs.h"
 #include "effect.h"
 #include "menu.h"
+#include "../widescreen.h"
 
 namespace ff7::battle
 {
@@ -1066,6 +1067,16 @@ namespace ff7::battle
                             v_top = page_spt_ptr->vScale * drawable->v_offset + drawable->v_offset / 2.0;
                             v_bottom = v_top_1 + v_top_2;
                         }
+
+                        // Temporary fix for Pollensalta cold breath bg widescreen fix 
+                        // (The correct solution should be to edit the file `magic/ff7/data/battle/special/hubuki/kemu.s` to edit the texture page)
+                        if(widescreen_enabled && (uint32_t)texture_ctx == ff7_externals.pollensalta_cold_breath_bg_texture_ctx)
+                        {
+                            float widescreen_multiplier = ((float)wide_viewport_width / (float)wide_viewport_height) / (4 / 3.f);
+                            quad_width *= widescreen_multiplier;
+                            quad_height *= widescreen_multiplier;
+                        }
+
                         drawable_state->vertices[0].x = x_left;
                         drawable_state->vertices[0].y = y_top;
                         drawable_state->vertices[0].z = 0.0;